repository layer with seeder done

+++ Task5_OnlineStore.DataAccess/Entities/Product.cs
@@ -11,7 +11,7 @@ namespace Task5_OnlineStore.DataAccess.Entities
         public int Id { get; set; }
         public string Name { get; set; }
         public string Description { get; set; }
-        public int Cost { get; set; }
+        public double Cost { get; set; }
         public int BrandId { get; set; }
         public Brand Brand { get; set; }
     }
